+++ changelog.md
@@ -20,4 +20,5 @@ The update script is able to migrate your old Hitbox streams to Smashcast. Howev
 - Lots of bug fixes
 - Backend styling improvements
 - Now users can click on table headers to sort table columns
+- Added support for these social links: Twitch, Instagram, Discord
 - Want more? Take a look: https://github.com/Contentify/Contentify/issues/264